Huge changes (first commit for a while =/)
Add in the nelder mead optimizer. This is basically defunct, though its usage
is commented out in swarm.cpp; look there to learn how to use it.
Add in the R-Tree. Should look at cleaning this code up a bit, but does a
pretty good job of the whole bounding rectangle thing.
Update stuff to be more c++11 (constexpr), as well as use the R-Tree.